# docs/swagger.yaml
basePath: /autotm-admin
definitions:
  dtos.V1BrandDTO:
    properties:
      id:
        type: integer
      logo_path:
        type: string
      name:
        type: string
    type: object
  dtos.V1BrandModelDTO:
    properties:
      brand_id:
        type: integer
      id:
        type: integer
      logo_path:
        type: string
      name:
        type: string
    type: object
  models.Brand:
    properties:
      id:
        type: integer
      logo_path:
        type: string
      name:
        type: string
    type: object
  models.BrandModel:
    properties:
      brand_id:
        type: integer
      brand_name:
        type: string
      id:
        type: integer
      logo_path:
        type: string
      name:
        type: string
    type: object
  models.BrandModelResult:
    properties:
      brand_models:
        items:
          $ref: '#/definitions/models.BrandModel'
        type: array
      count:
        type: integer
    type: object
  models.BrandResult:
    properties:
      brands:
        items:
          $ref: '#/definitions/models.Brand'
        type: array
      count:
        type: integer
    type: object
externalDocs:
  description: Other services
info:
  contact: {}
  description: AutoTM-Admin swagger
  title: AutoTM-Admin
  version: 3.0.0
paths:
  /brand/create-brand:
    post:
      consumes:
      - application/json
      description: Creates a new brand with the given name and logo path
      parameters:
      - description: Brand data
        in: body
        name: brand
        required: true
        schema:
          $ref: '#/definitions/dtos.V1BrandDTO'
      produces:
      - application/json
      responses:
        "200":
          description: Returns created brand ID
          schema:
            additionalProperties:
              type: integer
            type: object
        "400":
          description: Bad request
          schema:
            type: string
        "422":
          description: Unprocessable entity
          schema:
            type: string
        "500":
          description: Internal server error
          schema:
            type: string
      summary: Create a new brand
      tags:
      - Brand
  /brand/create-model:
    post:
      consumes:
      - application/json
      description: Creates a new brand model with the given name and logo path
      parameters:
      - description: Brand Model data
        in: body
        name: brand
        required: true
        schema:
          $ref: '#/definitions/dtos.V1BrandModelDTO'
      produces:
      - application/json
      responses:
        "200":
          description: Returns created model ID
          schema:
            additionalProperties:
              type: integer
            type: object
        "400":
          description: Bad request
          schema:
            type: string
        "422":
          description: Unprocessable entity
          schema:
            type: string
        "500":
          description: Internal server error
          schema:
            type: string
      summary: Create a new brand model
      tags:
      - Brand Model
  /brand/delete-brand:
    delete:
      consumes:
      - application/json
      description: Deletes a brand by ID
      parameters:
      - description: Brand ID to delete
        in: query
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Brand deleted successfully
          schema:
            type: string
        "400":
          description: Bad request
          schema:
            type: string
        "404":
          description: Brand not found
          schema:
            type: string
        "500":
          description: Internal server error
          schema:
            type: string
      summary: Delete a brand
      tags:
      - Brand
  /brand/delete-model:
    delete:
      consumes:
      - application/json
      description: Deletes a brand model by ID
      parameters:
      - description: Brand Model ID to delete
        in: query
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: brand model deleted successfully
          schema:
            type: string
        "400":
          description: Bad request
          schema:
            type: string
        "404":
          description: Brand Model not found
          schema:
            type: string
        "500":
          description: Internal server error
          schema:
            type: string
      summary: Delete a brand model
      tags:
      - Brand Model
  /brand/get-brands:
    get:
      consumes:
      - application/json
      description: Get a paginated list of brands with optional search
      parameters:
      - description: Limit number of brands to return
        in: query
        name: limit
        type: integer
      - description: Page number
        in: query
        name: page
        type: integer
      - description: Search string to filter brands by name
        in: query
        name: search
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: List of brands with pagination info
          schema:
            $ref: '#/definitions/models.BrandResult'
        "400":
          description: Bad request
          schema:
            type: string
        "500":
          description: Internal server error
          schema:
            type: string
      summary: Get all brands
      tags:
      - Brand
  /brand/get-models:
    get:
      consumes:
      - application/json
      description: Get a paginated list of brand models with optional search
      parameters:
      - description: Limit number of brand models to return
        in: query
        name: limit
        type: integer
      - description: Page number
        in: query
        name: page
        type: integer
      - description: Search string to filter brand models or brands by name
        in: query
        name: search
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: List of brand models with pagination info
          schema:
            $ref: '#/definitions/models.BrandModelResult'
        "400":
          description: Bad request
          schema:
            type: string
        "500":
          description: Internal server error
          schema:
            type: string
      summary: Get all brand models
      tags:
      - Brand Model
  /brand/update-brand:
    put:
      consumes:
      - application/json
      description: Updates brand details by ID
      parameters:
      - description: Brand data with ID
        in: body
        name: brand
        required: true
        schema:
          $ref: '#/definitions/dtos.V1BrandDTO'
      produces:
      - application/json
      responses:
        "200":
          description: Returns updated brand ID
          schema:
            additionalProperties:
              type: integer
            type: object
        "400":
          description: Bad request
          schema:
            type: string
        "422":
          description: Unprocessable entity
          schema:
            type: string
        "500":
          description: Internal server error
          schema:
            type: string
      summary: Update an existing brand
      tags:
      - Brand
  /brand/update-model:
    put:
      consumes:
      - application/json
      description: Updates brand model details by ID
      parameters:
      - description: Brand Model data with ID
        in: body
        name: brand
        required: true
        schema:
          $ref: '#/definitions/dtos.V1BrandModelDTO'
      produces:
      - application/json
      responses:
        "200":
          description: Returns updated brand model ID
          schema:
            additionalProperties:
              type: integer
            type: object
        "400":
          description: Bad request
          schema:
            type: string
        "422":
          description: Unprocessable entity
          schema:
            type: string
        "500":
          description: Internal server error
          schema:
            type: string
      summary: Update an existing brand model
      tags:
      - Brand Model
  /brand/upload-image:
    post:
      consumes:
      - multipart/form-data
      description: Uploads an image file and returns the file path
      parameters:
      - description: Image file to upload
        in: formData
        name: image
        required: true
        type: file
      produces:
      - application/json
      responses:
        "200":
          description: Returns the uploaded image path Successfully
        "400":
          description: Bad request
          schema:
            type: string
        "500":
          description: Internal server error
          schema:
            type: string
      summary: Upload an image
      tags:
      - Upload Image
schemes:
- http
- https
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: authorization
    type: apiKey
swagger: "2.0"